requirements.txt

aiohttp
ffmpeg
yt-dlp
python-dotenv